+++ CHANGES.md
@@ -8,6 +8,7 @@ Changes in 10.4
 - Add student name to Student's Absences and Grades delete prompt in MassDrops.php
 - Fix display only first letter of attendance code in AttendanceSummary.php
 - Remove "Minimum assignment points for letter grade" config option in Grades/Configuration.php & StudentGrades.php
+- Truncate Assignment title to 36 chars in StudentGrades.php
 
 Changes in 10.3.2
 -----------------